Make this starter IE11 ready
Our company values still advocate for a wide reach of users. On slow devices and even outdated browsers.
This PR fixes that.

I added some mechanisms to transpile node_modules, in case some `yarn installs` something that is not outright IE11 compatible, but sometimes that's not even enough.
Why? Because `wouter` wouldn't work even if it was transpiled. In that case, we replaced with `react-router-dom` which is heavier but more stable. It's the trade-off of supporting older browsers and total bundle size.

I also included polyfills which `babel` in turn turns to individual imports. So we don't need to import the whole damn thing. These polyfills are `core-js` and `regenerator-runtime`.