Remove CORS gem and configuration
We don't need CORS config as the is used in a single domain. If we need we should add CORS back in.

There is no problem is disabling this. It's best to keep the default restrictive behavior of Rails (disable all CORS requests) and when the problem arises we add this in.