Use Github Actions for CI (#400)
Add a Github Actions workflow for building and testing the project. Github Actions are 100% free for public repos, so it might a good idea to switch between this and Semaphore. The issues that plagued the early versions of Github Actions don't seem to exist anymore, and the build times are pretty acceptable now.

@filipebfernandes @LuisViana do we still want to keep the Semaphore config? I think we can remove that dependency entirely as Github Actions are cheaper and integrate better with Github (given that we are using Github itself). We can either delete the Semaphore config or just disable the project on Semaphore and keep the config around for projects where we don't have Github or just leave as is and test it on both places (GH and Semaphore). The choice is yours. Personally I'm all for spending less :joy: and using what Github offers, but your call on this.

Also updates Ruby and JS version to allow the new CI pipeline to work smoothly.
It also includes an update to `sassc` for it to work on a Github Actions environment.