Replace react-hot-loader with react-refresh
On this PR, I remove all traces of `react-hot-loader` and replace it with `react-refresh`, which is the new hot loading mechanism that the React team will support from now on. They brought that from React Native, which is pretty cool.

Dan Abramov himself (author of react-hot-loader) already signaled that the library eventually will die out, and this is the way to go now. I caught that on Twitter and replaced it on all the starter repos.