Upgrade ruby to 2.7.1 and nodejs to 12.19.0 (#598)
Ruby upgrade is a major release version. It had some deprecation warnings but most gems updated so now it's a good time for it. More performance and security improvements.

Only `sorcery`, the authentication gem, has a deprecation warning that is already fixed on their `master` branch. Once they upgrade it, dependabot will take care of that for us.